Rename files since we are actually supporting the standalone version
