+++ crates/cow-amm/src/implementations/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod safe_based;
+pub(crate) mod standalone;
